move out envs from repositories.go
